Remove columns prop from Table component
