+++ src/components/TableAntd/views/SortView/SortView.tsx
@@ -1,30 +1,39 @@
 import { useState } from 'react'
 
+import { ColumnsType } from 'antd/es/table'
+
 import { SearchSelect, SearchSelectProps } from '@banx/components/SearchSelect'
-import { SortDropdown, SortDropdownProps } from '@banx/components/SortDropdown'
+import { SortDropdown } from '@banx/components/SortDropdown'
 import { Toggle, ToggleProps } from '@banx/components/Toggle'
 
 import { ViewState, useTableView } from '@banx/store'
 
+import { SortParams } from '../../types'
 import { SwitchModeButtons } from './components'
+import { parseTableColumn } from './helpers'
 
 import styles from './SortView.module.less'
 
-interface SortViewProps<P> {
+interface SortViewProps<T, P> {
+  columns: ColumnsType<T>
   searchSelectParams: SearchSelectProps<P>
-  sortParams?: SortDropdownProps
+  sortParams?: SortParams
   toggleParams?: ToggleProps
   showCard?: boolean
 }
 
-const SortView = <P extends object>({
+const SortView = <T extends object, P extends object>({
+  columns,
   searchSelectParams,
   sortParams,
   toggleParams,
   showCard,
-}: SortViewProps<P>) => {
+}: SortViewProps<T, P>) => {
   const { viewState, setViewState } = useTableView()
   const [searchSelectCollapsed, setSearchSelectCollapsed] = useState(true)
+
+  const sortableColumns = columns.filter((column) => !!column.sorter)
+  const sortDropdownOptions = sortableColumns.map(parseTableColumn)
 
   const handleViewStateChange = (state: ViewState) => {
     setViewState(state)
@@ -41,7 +50,7 @@ const SortView = <P extends object>({
         <div className={styles.rowGap}>
           {showCard && <SwitchModeButtons viewState={viewState} onChange={handleViewStateChange} />}
           {toggleParams && <Toggle {...toggleParams} />}
-          {sortParams && <SortDropdown {...sortParams} />}
+          {sortParams && <SortDropdown options={sortDropdownOptions} {...sortParams} />}
         </div>
       )}
     </div>
